backend/requirements.txt
fastapi
uvicorn
pydantic
youtube-transcript-api
langchain-community
langchain-openai
faiss-cpu
tiktoken
python-dotenv
